Rename componentTypesView.ts to registriesView.ts
