+++ src/registriesView.ts
