+++ src/webview/devfile-registry/registryViewLoader.ts
@@ -9,7 +9,7 @@ import { ExtenisonID } from '../../util/constants';
 import { getInstance } from '../../odo';
 import { Command } from '../../odo/command';
 import { stringify } from 'yaml';
-import { ComponentTypesView } from '../../componentTypesView';
+import { ComponentTypesView } from '../../registriesView';
 import { StarterProject } from '../../odo/componentTypeDescription';
 import { DevfileComponentType } from '../../odo/componentType';
 import { vsCommand } from '../../vscommand';
